Improve league info cache load
Previously league info would load from cached API responses as long as the response was within the last day. This is still true, but added an additional case to load from cache if the league has already ended. In this case, we don't expect league info to update any more, so we can just keep using the cache.