Change API rate limiter
Previously the "rate limit" wasn't really a rate limit, more a concurrency limit with a semaphore that limited to 5 concurrent requests.

Now this uses an actual rate limit of 5 requests per second. In the future I may combine the two to account for having 5 simultaneous requests that are taking a while to finish.